#1306 - Get the new statutDonneesSource from the MTD's export XML
