+++ backend/geonature/core/gn_meta/routes.py
@@ -39,7 +39,7 @@ from geonature.core.gn_meta.repositories import (
 from utils_flask_sqla.response import json_resp
 from geonature.core.gn_permissions import decorators as permissions
 from geonature.core.gn_permissions.tools import cruved_scope_for_user_in_module
-from geonature.core.gn_meta import mtd_utils
+from geonature.core.gn_meta.mtd import mtd_utils
 import geonature.utils.filemanager as fm
